refactor(gatherer): refactor in order to unit-test it
In order to reduce and simplify the test cases in
trigger_controller_test.go, I decided to refactor the functions used by
the trigger controller, starting with DataForCertificate.

Note: DataForCertificate will now look for a certificate request with a
revision of "1" when the certificate has no revision set. This feature
is required so that we can "retry" immediately instead of waiting for 1
hour before retrying.

Signed-off-by: Maël Valais <mael.valais@gmail.com>