trigger_controller: now able to know when cert and CR are mismatched
Note that the wrong certificate request is given. The 'current'
certificate request returned by DataForCertificate corresponds to the
previous certificate request that led to the current certificate. But
what we really want is to compare the certificate's spec with the "next"
certificate request.

We probably need another return argument for DataForCertificate that
would return the "next" certificate request.

Signed-off-by: Maël Valais <mael@vls.dev>